BUG: Set/Get Covariance works for itkOrientedGaussianInterpolateImageFilter
Files needed to be reorganised. Covariance is now stored as FixedArray. In further consequence WrapITKTypes needed to be adapted so that WrapITK can understand that data format (that was also the reason why the previous version did not allow a "GetCovariance" for Python wrapping). Overall, files are now more conform with ITK coding style.

Next step: Implement averaged computation in neighbourhood of cindex.

Change-Id: Ibda89cfd3584b8fa6d7f5f30fb9cb1282b1694b9